Refatora fields para evitar repetição
