+++ transparenciagovbr/spiders/pagamento-hist/spiders.py
@@ -6,23 +6,9 @@ import zipfile
 import rows
 
 from transparenciagovbr import settings
+from transparenciagovbr.fields import BrazilianDateField, MoneyRealField
 from transparenciagovbr.spiders.base import TransparenciaBaseSpider
 from transparenciagovbr.utils.io import NotNullTextWrapper
-
-
-class BrazilianDateField(rows.fields.DateField):
-    INPUT_FORMAT = "%d/%m/%Y"
-
-
-class MoneyRealField(rows.fields.DecimalField):
-    @classmethod
-    def deserialize(cls, value):
-        """
-        >>> MoneyRealField.deserialize("89188,11")
-        '89188.11'
-        """
-        value = value.replace(",", ".")
-        return super().deserialize(value)
 
 
 SCHEMA_PATH = str(
