+++ transparenciagovbr/spiders/pagamento/spiders.py
@@ -3,23 +3,9 @@ import datetime
 import rows
 
 from transparenciagovbr import settings
+from transparenciagovbr.fields import BrazilianDateField, MoneyRealField
 from transparenciagovbr.spiders.base import TransparenciaBaseSpider
 from transparenciagovbr.utils.date import today
-
-
-class BrazilianDateField(rows.fields.DateField):
-    INPUT_FORMAT = "%d/%m/%Y"
-
-
-class MoneyRealField(rows.fields.DecimalField):
-    @classmethod
-    def deserialize(cls, value):
-        """
-        >>> MoneyRealField.deserialize("89188,11")
-        '89188.11'
-        """
-        value = value.replace(",", ".")
-        return super().deserialize(value)
 
 
 SCHEMA_PATH = str((settings.REPOSITORY_PATH / "schema" / "pagamento.csv").absolute())
